Statistical Singleton Class
We have added a Statistical Singleton class which uses the rest of the
DACs in the library to collect statistics about the website (number of
projects, number of contributions...).

What's more some methods have been added to the corresponding DACS and
the Master Page has been changed to use the singleton.

In addition, now the initial project creation contribution is inserted
as a normal contribution.

+++ ProjectShode/ProfileCredits.aspx.cs
@@ -25,7 +25,7 @@ namespace Project_Shode
         protected void addCredits(object sender, EventArgs e)
         {
             Button b = sender as Button;
-            float credits = float.Parse(b.CommandArgument.ToString());
+            int credits = Int32.Parse(b.CommandArgument.ToString());
 
             UserBE user = new UserBE();
             user.Nickname = Session["UserNickname"].ToString();
